Extract a concern for identifiers
Doesn't add any new code, just moves some methods into a concern
to make CocinaRecord smaller, since they are related.
